Initial proposal for TIP
This is the initial proposal for TIP (Technology Improvement Proposal).
A new idea that takes the ideas of PEP's from the Python community and
brings them into our organization. (PEP in Python's community stands for
"Python Enhancement Proposal")

Please read the README within the pull request for an idea of the proposal, its
initial ideas, and czars.

The czars mentioned in the README below do not have to be the eventual czars.
These are simply choices made by myself and Adam McKerlie. They were chosen for
their leadership within the organization and the contributions we feel they
could make to the process.

**It is important to note** that TIPs are just a proposal, just like this pull
request. We hope to bring unity to the department and welcome any changes you'd
like to contribute to this initial proposal

TIP 2 is linked to a `TODO`. The idea is we'll do a pull request for this
particular TIP, and fix the link.